Handle OS errors when storing data
When trying to access the storage directory to store new data items the
application may need to create a new file inside the directory or modify
the contents of an existing file. These operations may raise an OS
error. This errors are now handled correctly: the program logs the
error and exits with an error code.